move pydantic_v1 exception classes into core.plugins to be reusable
